+++ apps/skirmish/handlers/commands/transaction.py
@@ -9,7 +9,7 @@ from apps.skirmish.messages.events.transaction import WarriorDroppedSilver
 
 @message_registry.register_command(command=transaction.WarriorDropsSilver)
 def handle_warrior_drops_silver(*, context: transaction.WarriorDropsSilver) -> list[Event] | Event:
-    amount = int(round(max(random.gauss(10, 5), 0)))
+    amount = round(max(random.gauss(10, 5), 0))
 
     if amount > 0:
         return [
